refactor: 공통된 훅 재사용
